+++ geotrek/api/v2/serializers.py
@@ -661,6 +661,7 @@ if 'geotrek.trekking' in settings.INSTALLED_APPS:
         cities = serializers.SerializerMethodField()
         districts = serializers.SerializerMethodField()
         departure_city = serializers.SerializerMethodField()
+        labels = serializers.SerializerMethodField()
         web_links = WebLinkSerializer(many=True)
         view_points = HDViewPointSerializer(many=True)
 
@@ -769,6 +770,9 @@ if 'geotrek.trekking' in settings.INSTALLED_APPS:
 
         def get_districts(self, obj):
             return [district.pk for district in obj.published_districts]
+
+        def get_labels(self, obj):
+            return [label.pk for label in obj.published_labels]
 
         def get_departure_city(self, obj):
             geom = self.get_first_point(obj.geom)
@@ -1123,6 +1127,7 @@ if 'geotrek.outdoor' in settings.INSTALLED_APPS:
         pdf = serializers.SerializerMethodField('get_pdf_url')
         cities = serializers.SerializerMethodField()
         districts = serializers.SerializerMethodField()
+        labels = serializers.SerializerMethodField()
         web_links = WebLinkSerializer(many=True)
         view_points = HDViewPointSerializer(many=True)
 
@@ -1131,6 +1136,9 @@ if 'geotrek.outdoor' in settings.INSTALLED_APPS:
 
         def get_districts(self, obj):
             return [district.pk for district in obj.published_districts]
+
+        def get_labels(self, obj):
+            return [label.pk for label in obj.published_labels]
 
         def get_courses(self, obj):
             courses = []