+++ docs/changelog.rst
@@ -5,6 +5,10 @@ CHANGELOG
 2.99.0+dev (XXXX-XX-XX)
 -----------------------
 
+**New features**
+
+- Filter trek and outdoor site labels according to whether they are published or not (#3529)
+
 
 2.99.0 (2023-07-18)
 -----------------------